% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{compress}
\alias{compress}
\title{Return a compressed image}
\usage{
compress(img, n)
}
\arguments{
\item{img}{array}

\item{n}{integer}
}
\value{
array
}
\description{
Return a compressed image
}
\examples{
old_img <- array(1:24, dim = c(3, 4, 2))
(compressed_img <- compress(old_img))
}
