% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{Return a croped image of desired size by
removing pixels from borders of the image}
\usage{
crop(image, width, height)
}
\arguments{
\item{image}{input image as a array}

\item{width}{integer desired width for new image}

\item{height}{integer desired height for new image}
}
\value{
array for croped image
}
\description{
Return a croped image of desired size by
removing pixels from borders of the image
}
\examples{
old_img <- array(1:24, dim = c(10, 10, 3))
croped_img <- crop(old_img, 5, 5)
}
