% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpen.R
\name{sharpen}
\alias{sharpen}
\title{Detects and enhances the edges in the image and
returns a sharpened and monotoned version (the same size as the original).}
\usage{
sharpen(img)
}
\arguments{
\item{img}{array input image}
}
\value{
array sharpened monotone image
}
\description{
Detects and enhances the edges in the image and
returns a sharpened and monotoned version (the same size as the original).
}
\examples{
old_img <- array(1:(100 * 100 * 3), dim = c(100, 100, 3))
new_img <- sharpen(old_img)
}
