% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Return a shrinked image}
\usage{
shrink(img, n)
}
\arguments{
\item{img}{array}

\item{n}{integer}
}
\value{
array
}
\description{
Return a shrinked image
}
\examples{
old_img <- array(1:24, dim = c(3, 4, 2))
(shrinked_img <- shrink(old_img))
}
