% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpen.R
\name{sharpen}
\alias{sharpen}
\title{Return a sharpened image}
\usage{
sharpen(img)
}
\arguments{
\item{img}{array}
}
\value{
array
}
\description{
Return a sharpened image
}
\examples{
old_img <- array(1:(100 * 100 * 3), dim = c(100, 100, 3))
new_img <- sharpen(old_img)
}
