% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpen.R
\name{sharpen}
\alias{sharpen}
\title{Return a sharpened image}
\usage{
sharpen(img)
}
\arguments{
\item{img}{array}
}
\value{
array
}
\description{
Return a sharpened image
}
\examples{
old_img <- array(1:24, dim = c(3, 4, 2))
(new_img <- sharpen(old_img))
}
