% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Return a shrinked image of desired size}
\usage{
shrink(img, n)
}
\arguments{
\item{img}{input image as a array}

\item{width}{integer desired width for new image}

\item{height}{integer desired height for new image}
}
\value{
array for shrinked image
}
\description{
Return a shrinked image of desired size
}
\examples{
old_img <- array(1:24, dim = c(3, 4, 2))
(shrinked_img <- shrink(old_img))
}
