% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink.R
\name{shrink}
\alias{shrink}
\title{Return a shrinked image of desired size by
removing pixels from borders of the image}
\usage{
shrink(image, width, height)
}
\arguments{
\item{width}{integer desired width for new image}

\item{height}{integer desired height for new image}

\item{img}{input image as a array}
}
\value{
array for shrinked image
}
\description{
Return a shrinked image of desired size by
removing pixels from borders of the image
}
\examples{
old_img <- array(1:24, dim = c(10, 10, 3))
shrinked_img <- shrink(old_img, 5, 5)
}
